Consolidated agenthook
