--- frontend/src/components/ActionManager/index.tsx
@@ -1,3 +1,0 @@
-import ActionManager from "./ActionManager";
-
-export default ActionManager;
