+++ frontend/src/lib/hooks/useAppHook.ts
@@ -49,6 +49,22 @@ export default function useAppHook() {
         setIsMenuOpen(!isMenuOpen);
     }
 
+    function useIsMobile() {
+        const [isMobile, setIsMobile] = useState<boolean | undefined>(undefined)
+
+        useEffect(() => {
+            const mql = window.matchMedia(`(max-width: ${768 - 1}px)`)
+            const onChange = () => {
+            setIsMobile(window.innerWidth < 768)
+            }
+            mql.addEventListener("change", onChange)
+            setIsMobile(window.innerWidth < 768)
+            return () => mql.removeEventListener("change", onChange)
+        }, [])
+
+        return !!isMobile
+    }
+
     return {
         appVersion,
         isMobile,
@@ -58,6 +74,7 @@ export default function useAppHook() {
         loadingMessage,
         setLoadingMessage,
         isMenuOpen,
-        handleMenuOpen
+        handleMenuOpen,
+        useIsMobile
     }
 }
